Homebridge Link Development
Working on connecting Homebridge
(https://github.com/nfarina/homebridge) to My-Garage. Using this
generic http plugin:

https://www.npmjs.com/package/homebridge-http-garagedoorcontroller

To make it work so far I have the code on Homebridge plugin configured
and talking to  HB_Link.php. This code will call mgserver.php to
open/close doors. However, getStatus is not working correctly, so for
an interim test I have set up a status file for each door - status[0-2]
which will contain the status of the door. These files are updated when
you use either My-Garage or Homebridge to open the door. but not when
you manually open the door. Will look at the Monitor code to
temporarily address this issue.
So this is a WIP